requirements.txt
langchain-groq
langchain-openai
python-dotenv
langchain-community
langchain-chroma
langchain
pypdf
tiktoken
